requirements.txt
anthropic>=0.96.0
python-dotenv>=1.0.0
streamlit>=1.50.0
pandas>=2.0.0
pyarrow>=15.0.0
requests>=2.32.0